Moved creationg of SessionEventLoopGroup ouside of PostOffice
