+++ broker/src/main/java/io/moquette/broker/PostOffice.java
@@ -19,7 +19,6 @@ import io.moquette.interception.BrokerInterceptor;
 import io.moquette.broker.subscriptions.ISubscriptionsDirectory;
 import io.moquette.broker.subscriptions.Subscription;
 import io.moquette.broker.subscriptions.Topic;
-import io.moquette.interception.messages.InterceptExceptionMessage;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
 import io.netty.handler.codec.mqtt.*;
@@ -27,8 +26,17 @@ import io.netty.util.ReferenceCountUtil;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
-import java.util.concurrent.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
+import java.util.HashSet;
+import java.util.List;
+import java.util.Objects;
+import java.util.Set;
+import java.util.concurrent.Callable;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ConcurrentHashMap;
+import java.util.concurrent.ConcurrentMap;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
@@ -176,13 +184,14 @@ class PostOffice {
     private final SessionEventLoopGroup sessionLoops;
 
     PostOffice(ISubscriptionsDirectory subscriptions, IRetainedRepository retainedRepository,
-               SessionRegistry sessionRegistry, BrokerInterceptor interceptor, Authorizator authorizator, int sessionQueueSize) {
+               SessionRegistry sessionRegistry, BrokerInterceptor interceptor, Authorizator authorizator,
+               SessionEventLoopGroup sessionLoops) {
         this.authorizator = authorizator;
         this.subscriptions = subscriptions;
         this.retainedRepository = retainedRepository;
         this.sessionRegistry = sessionRegistry;
         this.interceptor = interceptor;
-        this.sessionLoops = new SessionEventLoopGroup(interceptor, sessionQueueSize);
+        this.sessionLoops = sessionLoops;
     }
 
     public void init(SessionRegistry sessionRegistry) {
@@ -611,97 +620,6 @@ class PostOffice {
 
     String sessionLoopThreadName(String clientId) {
         return sessionLoops.sessionLoopThreadName(clientId);
-    }
-
-    static class SessionEventLoopGroup {
-        private static final Logger LOG = LoggerFactory.getLogger(SessionEventLoopGroup.class);
-
-        private final SessionEventLoop[] sessionExecutors;
-        private final BlockingQueue<FutureTask<String>>[] sessionQueues;
-        private final int eventLoops = Runtime.getRuntime().availableProcessors();
-        private final ConcurrentMap<String, Throwable> loopThrownExceptions = new ConcurrentHashMap<>();
-
-        SessionEventLoopGroup(BrokerInterceptor interceptor, int sessionQueueSize) {
-            this.sessionQueues = new BlockingQueue[eventLoops];
-            for (int i = 0; i < eventLoops; i++) {
-                this.sessionQueues[i] = new ArrayBlockingQueue<>(sessionQueueSize);
-            }
-            this.sessionExecutors = new SessionEventLoop[eventLoops];
-            for (int i = 0; i < eventLoops; i++) {
-                SessionEventLoop newLoop = new SessionEventLoop(this.sessionQueues[i]);
-                newLoop.setName(sessionLoopName(i));
-                newLoop.setUncaughtExceptionHandler((loopThread, ex) -> {
-                    // executed in session loop thread
-                    // collect the exception thrown to later re-throw
-                    loopThrownExceptions.put(loopThread.getName(), ex);
-
-                    // This is done in asynch from another thread in BrokerInterceptor
-                    interceptor.notifyLoopException(new InterceptExceptionMessage(ex));
-                });
-                newLoop.start();
-                this.sessionExecutors[i] = newLoop;
-            }
-        }
-
-        int targetQueueOrdinal(String clientId) {
-            return Math.abs(clientId.hashCode()) % this.eventLoops;
-        }
-
-        private String sessionLoopName(int i) {
-            return "Session Executor " + i;
-        }
-
-        String sessionLoopThreadName(String clientId) {
-            final int targetQueueId = targetQueueOrdinal(clientId);
-            return sessionLoopName(targetQueueId);
-        }
-
-        /**
-         * Route the command to the owning SessionEventLoop
-         * */
-        public RouteResult routeCommand(String clientId, String actionDescription, Callable<String> action) {
-            SessionCommand cmd = new SessionCommand(clientId, action);
-            final int targetQueueId = targetQueueOrdinal(cmd.getSessionId());
-            LOG.debug("Routing cmd [{}] for session [{}] to event processor {}", actionDescription, cmd.getSessionId(), targetQueueId);
-            final FutureTask<String> task = new FutureTask<>(() -> {
-                cmd.execute();
-                cmd.complete();
-                return cmd.getSessionId();
-            });
-            if (Thread.currentThread() == sessionExecutors[targetQueueId]) {
-                SessionEventLoop.executeTask(task);
-                return RouteResult.success(clientId, cmd.completableFuture());
-            }
-            if (this.sessionQueues[targetQueueId].offer(task)) {
-                return RouteResult.success(clientId, cmd.completableFuture());
-            } else {
-                LOG.warn("Session command queue {} is full executing action {}", targetQueueId, actionDescription);
-                return RouteResult.failed(clientId);
-            }
-        }
-
-        public void terminate() {
-            for (SessionEventLoop processor : sessionExecutors) {
-                processor.interrupt();
-            }
-            for (SessionEventLoop processor : sessionExecutors) {
-                try {
-                    processor.join(5_000);
-                } catch (InterruptedException ex) {
-                    LOG.info("Interrupted while joining session event loop {}", processor.getName(), ex);
-                }
-            }
-
-            for (Map.Entry<String, Throwable> loopThrownExceptionEntry : loopThrownExceptions.entrySet()) {
-                String threadName = loopThrownExceptionEntry.getKey();
-                Throwable threadError = loopThrownExceptionEntry.getValue();
-                LOG.error("Session event loop {} terminated with error", threadName, threadError);
-            }
-        }
-
-        public int getEventLoopCount() {
-            return eventLoops;
-        }
     }
 
     /**
